+++ components/Button/index.tsx
@@ -5,12 +5,35 @@ import setup, { iButton } from './setup.ts';
  * @todo [!] Add a boolean property for smaller buttons.
  */
 export default function (props: Partial<iButton>) {
-  const { c, nostyle, nostyleAll, fref, children, compact, type, ...p } = setup(
+  const {
+    c,
+    nostyle,
+    nostyleAll,
+    fref,
+    children,
+    compact,
+    type,
+    OnExtendedClick,
+    ...p
+  } = setup(
     props,
   );
 
   return (
-    <button ref={fref} class={c.button} {...p}>
+    <button
+      onClick={OnExtendedClick}
+      onTouchEnd={OnExtendedClick}
+      onKeyUp={OnExtendedClick
+        ? (ev) => {
+          if (ev.code === 'Space') {
+            OnExtendedClick();
+          }
+        }
+        : undefined}
+      ref={fref}
+      class={c.button}
+      {...p}
+    >
       {children}
     </button>
   );