+++ components/Menu/setup.ts
@@ -31,10 +31,11 @@ const style = {
     margin: var(--s-quarter) 0;
     background-color: var(--clr-bg-panel-50);
     border-radius: var(--s-quarter);
-    backdrop-filter: blur(var(--s-three-quarters));
-    -webkit-backdrop-filter: blur(var(--s-three-quarters));
+    backdrop-filter: blur(var(--s-three-quarters)) brightness(75%);
+    -webkit-backdrop-filter: blur(var(--s-three-quarters)) brightness(75%);
     box-shadow: 0px var(--s-eighth) var(--s-quarter) var(--s-eighth) rgba(0, 0, 0, 0.1);
     position: absolute;
+    z-index: 1;
   `,
   container: css`
     display: flex;