+++ components/Navigation/setup.ts
@@ -30,7 +30,7 @@ const style = {
   wrapper: css`
     top: 0;
     width: 100%;
-    z-index: 1;
+    z-index: 20;
   `,
 };